+++ flutterx/lib/flutterx.dart
@@ -1,1 +1,5 @@
 library flutterx;
+
+import 'dart:ui';
+
+part 'src/color.dart';