+++ testx/lib/testx.dart
@@ -1,0 +1,5 @@
+library testx;
+
+import 'package:test/test.dart';
+
+part 'src/matcher/dynamic_matcher.dart';